# Improve selectRoomIDsWithAnyMembershipSQL performance